+++ src/rpc/client.cpp
@@ -75,7 +75,7 @@ static const CRPCConvertParam vRPCConvertParams[] = {
     {"listunspent", 1, "maxconf"},
     {"listunspent", 2, "addresses"},
     {"listunspent", 4, "query_options"},
-    {"getblock", 1, "verbose"},
+    {"getblock", 1, "verbosity"},
     {"getblockheader", 1, "verbose"},
     {"getchaintxstats", 0, "nblocks"},
     {"gettransaction", 1, "include_watchonly"},