+++ src/rpc/server.cpp
@@ -18,6 +18,8 @@
 #include <univalue.h>
 
 #include <boost/algorithm/string/case_conv.hpp> // for to_upper()
+#include <boost/algorithm/string/classification.hpp>
+#include <boost/algorithm/string/split.hpp>
 #include <boost/bind.hpp>
 #include <boost/signals2/signal.hpp>
 
@@ -461,8 +463,17 @@ transformNamedArguments(const JSONRPCRequest &in,
     }
     // Process expected parameters.
     int hole = 0;
-    for (const std::string &argName : argNames) {
-        auto fr = argsIn.find(argName);
+    for (const std::string &argNamePattern : argNames) {
+        std::vector<std::string> vargNames;
+        boost::algorithm::split(vargNames, argNamePattern,
+                                boost::algorithm::is_any_of("|"));
+        auto fr = argsIn.end();
+        for (const std::string &argName : vargNames) {
+            fr = argsIn.find(argName);
+            if (fr != argsIn.end()) {
+                break;
+            }
+        }
         if (fr != argsIn.end()) {
             for (int i = 0; i < hole; ++i) {
                 // Fill hole between specified parameters with JSON nulls, but